Rework of breath/suffocation mechanic
-breath is now stored as a byte representing a direct percentage
-the breath is decremented only on every second when in a non-breathable block and by 5% (meaning 20 seconds before taking damage - now decoupled from tick rate)
-when out of breath, suffocation does 10 points of health damage per second (meaning at most 10 seconds before death after running out of breath)

+++ october-common/src/main/java/com/jeffdisher/october/types/CreatureEntity.java
@@ -18,7 +18,7 @@ public record CreatureEntity(int id
 		// The health value of the entity.  Currently, we just use a byte since it is in the range of [1..100].
 		, byte health
 		// The breath the entity has (for drowning).
-		, int breath
+		, byte breath
 		// ----- Data elements below this line are considered ephemeral and will NOT be persisted. -----
 		// This data field is defined by helpers based on the type (remember that it is NOT persistent).
 		, Object extendedData
